feat: 리더 코드 수정

+++ src/main/java/com/blaybus/appsolute/leaderquest/domain/request/LeQuestBoardRequest.java
@@ -10,9 +10,10 @@ public class LeQuestBoardRequest {
 
     private Long userId;
     private Long leaderQuestId;
-    private Double actualPoint;
+    private Long actualPoint;
     public LeQuestBoard.QuestStatus questStatus;
     private String employeeNumber;
     private String leaderQuestName;
     private Long month;
+
 }
